Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/script-src-parsing-implicit-and-explicit-port-number-expected.txt

Tests script-src source expression matching with implicit and explicit default port numbers.



--------
Frame: '<!--frame1-->'
--------
PASS

--------
Frame: '<!--frame2-->'
--------
PASS

--------
Frame: '<!--frame3-->'
--------
PASS

--------
Frame: '<!--frame4-->'
--------
PASS

--------
Frame: '<!--frame5-->'
--------
PASS
